refactor(image): extract pure-Dart image decode into pdf_graphics
Move the dart:ui-free half of image-XObject decoding — Flate/CCITT/JBIG2/
JPX/CMYK-JPEG sample decode, _toRgba colour conversion, /Decode, color-key,
soft/stencil masks, ICC — out of dart_pdf_editor's image_decoder.dart and
into pdf_graphics/src/image_pixels.dart. This is the seam issue #73 needs:
the pure decode must be importable without dart:ui so a render worker (a
native isolate, and crucially a Web Worker with no Flutter engine) can run it
off the UI thread.

New public API in pdf_graphics:
- decodePdfImageBase -> PdfImageBase: base samples to straight RGBA, no mask.
- decodePdfImagePixels -> PdfDecodedPixels: premultiplied, codec-ready RGBA
  for everything but the platform JPEG codec (a non-CMYK DCTDecode base, or
  an image under a DCTDecode /SMask), where it returns null.
- helpers the dart:ui JPEG path reuses (pdfApplyImageAlpha,
  pdfApplyImageDecodeAndColorKey, pdfImageSoftMask/StencilMask,
  pdfImageDctSoftMaskBytes, pdfImage{DecodeRanges,ColorKeyRanges,ColorFamily,
  Filters}, pdfPremultiplyRgba).

image_decoder.dart is now the thin dart:ui executor: decodePdfImagePixels as
a fast path, decodePdfImageBase + the platform codec for a DCT mask, and the
platform JPEG-base path. decodeImages gains a `decoded` map so callers can
supply worker-decoded pixels and skip straight to decodeImageFromPixels.

Splitting the base out fixes the one case a naive "decline on DCT mask" got
wrong — a purely-decoded base (CMYK/Flate) under a JPEG /SMask, which would
otherwise mis-decode or vanish (caught by GWG166 in the Ghent baseline).

Pure-Dart decode pinned by pdf_graphics/test/image_pixels_test.dart; all
image/render suites (image_decoder, image_cache, record/replay, Ghent
+40 -14 unchanged, pdf.js, corpus) green. No behavior change.
