+++ packages/pdf_graphics/lib/pdf_graphics.dart
@@ -12,6 +12,7 @@ export 'src/device.dart';
 export 'src/font_info.dart';
 export 'src/function.dart';
 export 'src/icc.dart';
+export 'src/image_pixels.dart';
 export 'src/interpreter.dart';
 export 'src/shading.dart';
 export 'src/matrix.dart';
