+++ packages/pdf_graphics/pubspec.yaml
@@ -18,6 +18,7 @@ environment:
 dependencies:
   pdf_cos: ^1.0.0
   pdf_document: ^1.0.0
+  image: ^4.9.1
 
 dev_dependencies:
   pdf_test_fixtures: ^1.0.0
